+++ docs/news.rst
@@ -4,6 +4,8 @@ Release Notes
 1.11.0 (unreleased)
 -------------------
 
+* Add Chrome and Firefox options to capabilities for remote servers.
+
 * Add :code:`firefox_arguments` and :code:`firefox_preferences` markers to
   specify arguments and preferences to pass to the :code:`firefox_options`
   fixture. Run :code:`pytest --markers` for details.